pyproject.toml
[project]
name = "discordsupportbot-cs"
version = "0.1.0"
description = "Member support chatbot for IEEE CS Discord server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-cerebras>=0.5.0",
    "langgraph>=0.6.6",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
